+++ api-demo/assets/demo-input.ts
@@ -6,6 +6,9 @@
    * This method is part of the {@link core-library/Math | Math subsystem}.
    * This incomplete HTML tag should be reported as an error: <tag
    *
+   * @privateRemarks
+   * This content should @not show up on the web site.
+   *
    * @param x - The first number to add
    * @param y - The second number to add
    * @returns The sum of `x` and `y`