+++ tsdoc/src/details/StandardModifierTagSet.ts
@@ -10,76 +10,76 @@ export class StandardModifierTagSet extends ModifierTagSet {
    * Returns true if the `@alpha` modifier tag was specified.
    */
   public isAlpha(): boolean {
-    return this.hasModifierTagWithUpperCase(StandardTags.alpha.tagNameWithUpperCase);
+    return this.hasTag(StandardTags.alpha);
   }
 
   /**
    * Returns true if the `@beta` modifier tag was specified.
    */
   public isBeta(): boolean {
-    return this.hasModifierTagWithUpperCase(StandardTags.beta.tagNameWithUpperCase);
+    return this.hasTag(StandardTags.beta);
   }
 
   /**
    * Returns true if the `@eventProperty` modifier tag was specified.
    */
   public isEventProperty(): boolean {
-    return this.hasModifierTagWithUpperCase(StandardTags.eventProperty.tagNameWithUpperCase);
+    return this.hasTag(StandardTags.eventProperty);
   }
 
   /**
    * Returns true if the `@experimental` modifier tag was specified.
    */
   public isExperimental(): boolean {
-    return this.hasModifierTagWithUpperCase(StandardTags.experimental.tagNameWithUpperCase);
+    return this.hasTag(StandardTags.experimental);
   }
 
   /**
    * Returns true if the `@internal` modifier tag was specified.
    */
   public isInternal(): boolean {
-    return this.hasModifierTagWithUpperCase(StandardTags.internal.tagNameWithUpperCase);
+    return this.hasTag(StandardTags.internal);
   }
 
   /**
    * Returns true if the `@override` modifier tag was specified.
    */
   public isOverride(): boolean {
-    return this.hasModifierTagWithUpperCase(StandardTags.override.tagNameWithUpperCase);
+    return this.hasTag(StandardTags.override);
   }
 
   /**
    * Returns true if the `@packageDocumentation` modifier tag was specified.
    */
   public isPackageDocumentation(): boolean {
-    return this.hasModifierTagWithUpperCase(StandardTags.packageDocumentation.tagNameWithUpperCase);
+    return this.hasTag(StandardTags.packageDocumentation);
   }
 
   /**
    * Returns true if the `@public` modifier tag was specified.
    */
   public isPublic(): boolean {
-    return this.hasModifierTagWithUpperCase(StandardTags.public.tagNameWithUpperCase);
+    return this.hasTag(StandardTags.public);
   }
 
   /**
    * Returns true if the `@readonly` modifier tag was specified.
    */
   public isReadonly(): boolean {
-    return this.hasModifierTagWithUpperCase(StandardTags.readonly.tagNameWithUpperCase);
+    return this.hasTag(StandardTags.readonly);
   }
 
   /**
    * Returns true if the `@sealed` modifier tag was specified.
    */
   public isSealed(): boolean {
-    return this.hasModifierTagWithUpperCase(StandardTags.sealed.tagNameWithUpperCase);
+    return this.hasTag(StandardTags.sealed);
   }
 
   /**
    * Returns true if the `@virtual` modifier tag was specified.
    */
   public isVirtual(): boolean {
-    return this.hasModifierTagWithUpperCase(StandardTags.virtual.tagNameWithUpperCase);
+    return this.hasTag(StandardTags.virtual);
   }
 }